+++ server/src/main/kotlin/org/elaastic/auth/UserLink.kt
@@ -1,4 +1,4 @@
-package org.elaastic.auth.cas
+package org.elaastic.auth
 
 import org.elaastic.common.persistence.AbstractJpaPersistable
 import org.elaastic.user.User
